+++ src/Handler/SyslogHandlerFactory.php
@@ -11,11 +11,11 @@ class SyslogHandlerFactory implements FactoryInterface
 {
     public function __invoke(array $options)
     {
-        $ident = (string) $options['ident'] ?? '';
-        $facility = (int) $options['facility'] ?? LOG_USER;
-        $level = $options['level'] ?? Logger::DEBUG;
-        $bubble = (boolean) $options['bubble'] ?? true;
-        $logOpts = $options['logOpts'] ?? LOG_PID;
+        $ident    = (string)  ($options['ident']    ?? '');
+        $facility = (int)     ($options['facility'] ?? LOG_USER);
+        $level    = (int)     ($options['level']    ?? Logger::DEBUG);
+        $bubble   = (boolean) ($options['bubble']   ?? true);
+        $logOpts  = (int)     ($options['logOpts']  ?? LOG_PID);
 
         return new SyslogHandler($ident, $facility, $level, $bubble, $logOpts);
     }
